refactor(dip): switch Handler to Port interfaces, reduce pub API surface by 38.4%
- Create 8 new Port interfaces (ProtocolMetricsPort, OffsetManagerPort, etc.)
- ISP split: BrokerRegistryPort (query) + BrokerLifecyclePort (lifecycle)
- Convert 624 pub fn -> fn across service/infra layers
- Remove -enable-globals from Makefile
- Clean Architecture: handler.v concrete imports 8 -> 0

Closes codebase-health-v3

+++ src/infra/protocol/kafka/handler_fetch.v
@@ -524,7 +524,7 @@ fn (mut h Handler) compress_records_for_fetch(records_data []u8, compression_typ
 	}
 
 	// Perform compression
-	compressed := h.compression_service.compress(records_data, compression_type) or {
+	compressed := h.compression_service.compress(records_data, i16(compression_type)) or {
 		h.logger.warn('Compression failed, returning uncompressed data', observability.field_string('topic',
 			topic_name), observability.field_int('partition', partition), observability.field_string('compression_type',
 			compression_type.str()), observability.field_err_str(err.str()))
